Reorganizacao dos arquivos para modularizar
