+++ funcoes.h
@@ -1,0 +1,14 @@
+#ifndef FUNCOES_H
+#define FUNCOES_H
+
+int segundo();
+int minuto();
+int hora();
+void unidade_tempo();
+
+// 1 - Função de Conversão de Unidade de Volume
+double obterFatorParaLitros(int unidade);
+int exibirSelecaoUnidade(const char *titulo);
+void converterVolume();
+
+#endif
