+++ main.c
@@ -1,1 +1,9 @@
+#include <stdio.h>
+#include "funcoes.h"
 
+int main()
+{
+  unidade_tempo();
+  converterVolume();
+  return 0;
+}
